requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
discord.py==2.5.2
distlib==0.3.9
filelock==3.16.1
Flask==3.1.0
fonttools==4.57.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.3.0
huggingface-hub==0.29.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kaleido==0.2.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.4.3
multitasking==0.0.11
narwhals==1.34.1
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.2.1
platformdirs==4.3.6
plotly==6.0.1
propcache==0.3.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.15.2
six==1.17.0
soupsieve==2.6
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
virtualenv==20.28.0
Werkzeug==3.1.3
yarl==1.19.0
yfinance==0.2.55
